services/requirements.txt
gevent
ari
odoorpc
py-Asterisk
git+https://github.com/litnimax/PyAjam.git
requests
setproctitle
terminado
